US - 06 - Standings Management: CRUD for standings with its DTOs, mapper and some adjustments to Match and MatchService

+++ src/main/java/com/kicktime/backend/repository/FieldRepository.java
@@ -3,5 +3,5 @@ package com.kicktime.backend.repository;
 import com.kicktime.backend.domain.model.Field;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FieldRepository extends JpaRepository<Field, Integer> {
+public interface FieldRepository extends JpaRepository<Field, Long> {
 }
